moved generate email verification token to AuthService from UserService.
