+++ src/main/java/com/roadmaps/Roadmaps/modules/user/events/listener/UserRegistrationEventListener.java
@@ -1,8 +1,8 @@
 package com.roadmaps.Roadmaps.modules.user.events.listener;
 
+import com.roadmaps.Roadmaps.modules.authentication.service.AuthService;
 import com.roadmaps.Roadmaps.modules.email.service.EmailService;
 import com.roadmaps.Roadmaps.modules.user.events.UserRegistrationEvent;
-import com.roadmaps.Roadmaps.modules.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class UserRegistrationEventListener {
-    private final UserService userService;
+    private final AuthService authService;
     private final EmailService emailService;
 
     @EventListener
@@ -21,7 +21,7 @@ public class UserRegistrationEventListener {
     public void handleUserRegistrationEvent(UserRegistrationEvent userRegistrationEvent) {
         String userEmail = userRegistrationEvent.getUser().getEmail();
         String frontendBaseUrl = userRegistrationEvent.getFrontendBaseUrl();
-        String verificationToken = userService.generateEmailVerificationToken(userRegistrationEvent.getUser());
+        String verificationToken = authService.generateEmailVerificationToken(userRegistrationEvent.getUser());
         if(verificationToken != null){
             emailService.sendVerificationEmailAsync(userEmail, verificationToken, frontendBaseUrl );
         } else {
